requirements.txt
acme==1.21.0
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.8.0
bcrypt==3.2.0
black==21.10b0
blinker==1.6.2
certbot==1.21.0
certbot-nginx==1.21.0
certifi==2020.6.20
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.4
ConfigArgParse==1.5.3
configobj==5.0.6
cryptography==3.4.8
dbus-python==1.2.18
Deprecated==1.2.14
distlib==0.3.4
distro==1.7.0
Django==5.1
djangorestframework==3.15.2
dlib==19.24.4
dnspython==2.6.1
duplicity==2.2.4
example-package==0.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
fasteners==0.14.1
filelock==3.6.0
Flask==3.0.3
Flask-Babel==2.0.0
Flask-Cors==3.0.8
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
flux==1.3.5
frozenlist==1.4.0
future==0.18.2
geographiclib==2.0
geopy==2.4.0
greenlet==2.0.2
gunicorn==22.0.0
hiredis==2.2.3
httplib2==0.20.2
idna==2.6
ImageHash==4.3.1
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.4.2
josepy==1.10.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.3
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
mypy-extensions==0.4.3
netifaces==0.11.0
numpy==2.0.0
oauthlib==3.2.0
olefile==0.46
opencv-python==4.10.0.84
package_name==0.1
packaging==24.1
parameterized==0.9.0
paramiko==2.9.3
parsedatetime==2.6
pathspec==0.9.0
pep8==1.7.1
pexpect==4.8.0
Pillow==9.0.1
pipenv==11.9.0
platformdirs==2.5.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.6.0
pycparser==2.22
pycryptodome==3.19.0
pycups==2.0.1
pydantic==2.9.2
pydantic_core==2.23.4
pygame==2.6.1
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.8.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-dateutil==2.8.1
python-decouple==3.8
pytz==2022.1
PyWavelets==1.6.0
pyxdg==0.27
PyYAML==5.4.1
redis==5.0.0
reportlab==3.6.8
requests==2.32.3
requests-toolbelt==0.9.1
scikit-learn==1.5.1
scipy==1.14.0
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
somepackage==1.2.3
SQLAlchemy==2.0.22
sqlparse==0.4.2
systemd-python==234
telebirr==0.1.11
threadpoolctl==3.5.0
tomli==1.2.2
twilio==8.8.0
typing_extensions==4.12.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.4
usb-creator==0.3.7
virtualenv==20.13.0+ds
virtualenv-clone==0.3.0
vonage==3.17.1
vonage-jwt==1.1.1
vonage-utils==1.1.2
wadllib==1.3.6
Werkzeug==3.0.3
whitenoise==6.7.0
wrapt==1.16.0
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
